% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bmi_bsa.R
\name{calc_bmi_bsa}
\alias{calc_bmi_bsa}
\title{Calculate BMI and BSA}
\usage{
calc_bmi_bsa(data = NULL)
}
\arguments{
\item{data}{A data frame with columns of \code{HTCM} and \code{WTKG}.}
}
\value{
A data frame with columns of \code{BMI}, \code{BSA1}, \code{BSA2}, and \code{BSA3} added
to \code{data}:
\itemize{
\item \code{BMI}: Body mass index in kilograms per meter squared, rounded to 1
decimal place.
\item \code{BSA1}: Body surface area in meters squared, rounded to 2 decimal
places; calculated using the Mosteller equation.
\item \code{BSA2}: Body surface area in meters squared, rounded to 2 decimal
places; calculated using the Gehan and George equation.
\item \code{BSA3}: Body surface area in meters squared, rounded to 2 decimal
places; calculated using the DuBois equation.
}
}
\description{
BMI and BSA are calculated for a data frame that minimally has columns of
HTCM and WTKG. Output columns match the definitions given by \code{sim_kid()}.
}
\examples{
demo0 <- sim_kid()
demo <- calc_bmi_bsa(data = demo0)
}
