% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_kid_nperagebin.R
\name{sim_kid_nperagebin}
\alias{sim_kid_nperagebin}
\title{Create body size metrics for virtual subjects using anthropometric growth chart data and an equal number of subjects per age bin}
\usage{
sim_kid_nperagebin(
  num_per_age = 1,
  prob_female = 0.5,
  age0isbirth = FALSE,
  age0to2yr_growthchart = "CDC",
  age2to20yr_correlate_htwt = TRUE,
  htwt_percentile_min = NULL,
  htwt_percentile_max = NULL,
  masterseed = NULL
)
}
\arguments{
\item{num_per_age}{A positive integer that specifies the number of subjects to simulate for each age bin. Defaults to a single subject for each age bin.}

\item{prob_female}{A numeric value with an inclusive range of \code{0} to \code{1} that specifies the probability that a given virtual subject is female. Defaults to \code{0.5}.}

\item{age0isbirth}{A logical that specifies whether age equal to zero denotes birth.
\itemize{
\item \code{TRUE}: Age of \code{0} is birth.
\item \code{FALSE} (the default): Age of \code{0} is ages from birth to less than one month.
\item Not applicable nor used for \code{age0to2yr_growthchart = "FENTON"}, for which postnatal age is always zero.
}}

\item{age0to2yr_growthchart}{A string that specifies which anthropometric growth charts are used for ages less than or equal to 2 years old.
\itemize{
\item \code{"CDC"} (the default): United States Centers for Disease Control and Prevention growth charts are used.
\item \code{"WHO"}: World Health Organization growth charts are used.
\item \code{"FENTON"}: Fenton growth charts for preterm newborns are used. This option is only available when simulating virtual subjects at birth (postnatal age = 0).
}}

\item{age2to20yr_correlate_htwt}{A logical that specifies whether correlations, by sex and year of age, are implemented between simulated height and simulated weight for ages greater than or equal to 2 years old.
\itemize{
\item \code{TRUE} (the default): Correlations are implemented between simulated height and simulated weight according to an identical internal-systems-data version of \link{cdc_ages2to20yr_correlations_by_sex_htcm_wtkg_summarized} located within the \code{data} folder.
\item \code{FALSE}: Height and weight are simulated independently without any correlation(s). Note that this will likely result in unrealistic virtual subjects.
\item Not applicable nor used for \code{age0to2yr_growthchart = "FENTON"}, for which height is not simulated.
}}

\item{htwt_percentile_min}{A numeric value that specifies the minimum allowed percentile of simulated height and weight, expressed as a decimal.
\itemize{
\item Must be greater than or equal to \code{0.001}.
\item Must be less than to \code{htwt_percentile_max}.
\item Defaults to \code{0.001} when \code{age0to2yr_growthchart = "CDC"} or \code{age0to2yr_growthchart = "WHO"}.
\item Defaults to \code{0.01} when \code{age0to2yr_growthchart = "FENTON"} to avoid non-viable birth weights.
}}

\item{htwt_percentile_max}{A numeric value that specifies the maximum allowed percentile of simulated height and weight, expressed as a decimal.
\itemize{
\item Must be greater than \code{htwt_percentile_min}.
\item Must be less than or equal to \code{0.999}.
\item Defaults to \code{0.999} when \code{age0to2yr_growthchart = "CDC"} or \code{age0to2yr_growthchart = "WHO"}.
\item Defaults to \code{0.99} when \code{age0to2yr_growthchart = "FENTON"} to avoid non-viable birth weights.
}}

\item{masterseed}{An integer ranging from \code{1} to \code{.Machine$integer.max} that sets an overall seed for the simulation to ensure reproducibility of the results. Defaults to no seed.}
}
\value{
A data frame with the number of rows equal to \code{num_per_age} multiplied by the number of anthropometric growth chart age bins and columns of:
\itemize{
\item \code{ID}: An integer ranging from \code{1} to \code{num} that serves as a virtual subject identifier.
\item \code{SEXF}: An integer of value \code{0} for male or \code{1} for female.
\item \code{AGEMO}: Postnatal age in months.
\item \code{AGE}: Postnatal age in years.
\item \code{GAWK}: Gestational age in weeks.
\item \code{WTKG}: Body weight in kilograms, rounded to 2 decimal places.
\item \code{HTCM}: Body height in centimeters, rounded to the nearest centimeter.
\item \code{BMI}: Body mass index in kilograms per meter squared, rounded to 1 decimal place.
\item \code{BSA1}: Body surface area in meters squared, rounded to 2 decimal places; calculated using the Mosteller equation.
\item \code{BSA2}: Body surface area in meters squared, rounded to 2 decimal places; calculated using the Gehan and George equation.
\item \code{BSA3}: Body surface area in meters squared, rounded to 2 decimal places; calculated using the DuBois equation.
\item \code{ZWTKG}: The z-score of weight-for-height for ages 0 to 2 years, weight-for-age for ages greater than 2 years, and weight-for-gestational-age for newborns when using Fenton growth charts.
\item \code{ZHTCM}: The z-score of height-for-age.
\item \code{PWTKG}: The percentile of weight corresponding to the respective z-score.
\item \code{PHTCM}: The percentile of height corresponding to the respective z-score.
A warning will be returned if the simulation to create virtual subjects fails.
}
}
\description{
Body size metrics (height, weight, BMI, and BSA) are created for a population of virtual subjects.
The body size metrics reflect the anthropometric growth chart distribution(s) and correlations (ex. height vs weight) according to virtual subject age and sex.
The probability that a given subject is female are specified by the user. The simulation produces and equal number of virtual subjects for each anthropometric age bin.
For ages greater than 2 years, CDC growth charts are used.
For ages birth to 2 years, either CDC (the default) or WHO growth charts can be used. Note that while CDC growth charts are used to prevent a jump discontinuity at 2 years, WHO growth charts are recommended for ages 0 to 2 years.
For birth only (postnatal age of zero), Fenton growth charts for preterm can be used according to a distribution of gestational age.
Note that when using Fenton growth charts, only body weight will be simulated.
}
\examples{
# Simulate 1 subject per age bin for ages 0 to 20 years using
#   CDC growth charts and a 50\% probability of female sex.
df_kids <- sim_kid_nperagebin()
}
\seealso{
\code{\link[=sim_kid]{sim_kid()}} to simulate virtual subjects using a distribution of age.
}
