% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimKid-package.R
\docType{package}
\name{SimKid-package}
\alias{SimKid}
\alias{SimKid-package}
\title{SimKid: Simulate Virtual Pediatrics using Anthropometric Growth Charts}
\description{
Simulate a virtual population of subjects that has demographic distributions (height, weight, and BMI) and correlations (height and weight), by sex and age, which mimic those reported in real-world anthropometric growth charts (CDC, WHO, or Fenton).
}
\author{
\strong{Maintainer}: Andrew Santulli \email{asantulli@epd-llc.com}

Other contributors:
\itemize{
  \item Enhanced Pharmacodynamics LLC [copyright holder, funder]
}

}
\keyword{internal}
